+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [0.5.0] - 2026-08-27
+
+### Added
+
+- Add PostgreSQL compilation with numbered parameters.
+- Add named Collector-managed database connection selection to the Rust client and CLI.
+
 ## [0.4.3] - 2026-08-26
 
 ### Fixed